% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{assign_label}
\alias{assign_label}
\title{A function to assign labels to a data frame}
\usage{
assign_label(data, var = names(data), label = names(data))
}
\arguments{
\item{data}{a data frame}

\item{var}{the variables to assign labels}

\item{label}{the labels to be assigned}
}
\value{
a data frame with labels updated
}
\description{
A function to assign labels to a data frame
}
\examples{
metalite:::assign_label(r2rtf::r2rtf_adae)
metalite:::assign_label(r2rtf::r2rtf_adae, var = "USUBJID", label = "Unique subject id")
}
