% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{check_duplicate_name}
\alias{check_duplicate_name}
\title{Check duplicate name}
\usage{
check_duplicate_name(x)
}
\arguments{
\item{x}{a list}
}
\description{
Check duplicate name
}
\examples{
\dontrun{
check_duplicate_name(list(a = 1, b = 2, b = 3))
}

}
