% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_adam_mapping}
\alias{collect_adam_mapping}
\title{Collect \code{adam_mapping} from \code{meta_adam} by \code{name}}
\usage{
collect_adam_mapping(meta, name)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{name}{a value of keyword}
}
\description{
Collect \code{adam_mapping} from \code{meta_adam} by \code{name}
}
\examples{
library(r2rtf)
meta <- meta_dummy()

collect_adam_mapping(meta, "apat")
}
