% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_observation_index}
\alias{collect_observation_index}
\title{Collect observation record index from observation dataset}
\usage{
collect_observation_index(meta, population, observation, parameter)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}
}
\value{
a vector of patient index within the observation group
}
\description{
Collect observation record index from observation dataset
}
\examples{
library(r2rtf)
meta <- meta_dummy()

collect_observation_index(meta, "apat", "wk12", "ser")
}
