% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_observation_record}
\alias{collect_observation_record}
\title{Collect observation record from observation dataset}
\usage{
collect_observation_record(
  meta,
  population,
  observation,
  parameter,
  var = NULL
)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}

\item{var}{a character vector of additional variables to be displayed in the output.}
}
\value{
a data.frame showing the observation datset
}
\description{
The key variables used in \code{id}, \code{group}, and \code{subset} are displayed by default.
}
\examples{
library(r2rtf)
meta <- meta_dummy()

collect_observation_record(meta, "apat", "wk12", "ser")
collect_observation_record(meta, "apat", "wk12", "ser", var = "AEDECOD")
}
