% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_population}
\alias{collect_population}
\title{Collect specification for population definition}
\usage{
collect_population(meta, population, observation = NULL, parameter = NULL)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}
}
\value{
a list covering the filter of population,
observation (if input) and parameter (if input)
}
\description{
Collect specification for population definition
}
\examples{
library(r2rtf)
meta <- meta_dummy()
collect_population(meta, "apat")
collect_population(meta, "apat", "wk12")
collect_population(meta, "apat", "wk12", "ser")

}
