% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_population_id}
\alias{collect_population_id}
\title{Collect subject id information from population dataset}
\usage{
collect_population_id(meta, population)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}
}
\description{
Collect subject id information from population dataset
}
\examples{
library(r2rtf)
library(magrittr)
meta <- meta_dummy()

head(collect_population_id(meta, "apat"))
}
