% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_population_record}
\alias{collect_population_record}
\title{Collect population record from population dataset}
\usage{
collect_population_record(meta, population, var = NULL)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{var}{a character vector of additional variables to be displayed in the output.}
}
\description{
The key variables used in \code{id}, \code{group}, and \code{subset} are displayed by default.
}
\examples{
library(r2rtf)
library(magrittr)
meta <- meta_dummy()

head(collect_population_record(meta, "apat"))
head(collect_population_record(meta, "apat", var = "AGE"))

}
