% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_title}
\alias{collect_title}
\title{Collect specification for title}
\usage{
collect_title(meta, population, observation, parameter, analysis)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}

\item{analysis}{a character value of analysis term name.
The term name is used as key to link information.}
}
\value{
a vector of strings to compose the table captions
}
\description{
Collect specification for title
}
\examples{
library(r2rtf)
meta <- meta_dummy()

collect_title(meta, "apat", "wk12", "ser", "ae_summary")
collect_title(meta, "apat", "wk12", "ser", "ae_specific")
}
