% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_parameter}
\alias{define_parameter}
\title{Define analysis parameter meta information for ADaM dataset}
\usage{
define_parameter(meta, name, subset, ...)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{name}{a character value of term name. The term name is used as key to link information.}

\item{subset}{an expression to identify analysis records. Refer \code{subset()} function.}

\item{...}{additional variables.}
}
\description{
Define analysis parameter meta information for ADaM dataset
}
\examples{
meta_adam(
  population = r2rtf_adsl,
  observation = r2rtf_adae
) |>
  define_plan(plan = plan) |>
  define_parameter(
    name = "rel",
    subset = AEREL \%in\% c("POSSIBLE", "PROBABLE")
  )
}
