% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fmt_quote}
\alias{fmt_quote}
\title{Format quote}
\usage{
fmt_quote(x)
}
\arguments{
\item{x}{a vector of character string.}
}
\description{
Format quote
}
\examples{
\dontrun{
'a = "b"'
fmt_quote('a = "b"')
}
}
