% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{fmt_sentence}
\alias{fmt_sentence}
\title{Format sentence}
\usage{
fmt_sentence(x)
}
\arguments{
\item{x}{a vector of character string.}
}
\description{
Format sentence
}
\examples{
\dontrun{
  fmt_sentence(" a sentence that  needs to be cleaned  ")
}
}
