% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{get_label}
\alias{get_label}
\title{A function to get the labels of data frame columns}
\usage{
get_label(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Labels of the input data frame.
}
\description{
A function to get the labels of data frame columns
}
\examples{
get_label(r2rtf::r2rtf_adae)
}
