% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_add_total.R
\name{meta_add_total}
\alias{meta_add_total}
\title{Add duplicate data to enable a total group}
\usage{
meta_add_total(meta, total = "Total")
}
\arguments{
\item{meta}{A metalite object.}

\item{total}{A character value of total group name.}
}
\value{
a metadata with a group total added
}
\description{
Add duplicate data to enable a total group
}
\examples{
x <- meta_add_total(meta_example())
# check there is a new group called Total
table(x$data_population$TRTA)

}
