% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_dummy.R
\name{meta_dummy}
\alias{meta_dummy}
\title{Create a dummy \code{meta_adam} object}
\usage{
meta_dummy()
}
\value{
A metadata object.
}
\description{
This function is only for illustration purposes.
The \code{r2rtf} package is required.
}
\examples{
meta_dummy()
}
