% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_dummy_exploration.R
\name{meta_dummy_exploration}
\alias{meta_dummy_exploration}
\title{Create a data exploration \code{meta_adam} object}
\usage{
meta_dummy_exploration(
  data,
  group,
  name = "ase",
  subset = NULL,
  label = "All Subjects Enrolled"
)
}
\arguments{
\item{data}{a data frame}

\item{group}{a character value of group variable name in \code{data}.}
}
\description{
Create a data exploration \code{meta_adam} object
}
