% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_example.R
\name{meta_example}
\alias{meta_example}
\title{Create an example \code{meta_adam} object}
\usage{
meta_example()
}
\value{
A metadata object.
}
\description{
This function is only for illustration purposes.
The r2rtf package is required.
}
\examples{
meta_example()
}
