% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_inherit.R
\name{meta_inherit}
\alias{meta_inherit}
\title{Inherit meta information by keywords}
\usage{
meta_inherit(meta, inherit, name, overwrite = FALSE)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{inherit}{A \code{meta_adam} object to be inherit.}

\item{name}{A vector of keywords from \code{meta_inherit} to \code{meta_adam}.}

\item{overwrite}{A logical value to force mapping update.}
}
\value{
A metadata object with population defined.
}
\description{
Inherit meta information by keywords
}
\examples{
meta_adam(
  population = r2rtf::r2rtf_adsl,
  observation = r2rtf::r2rtf_adae
) |>
  meta_inherit(meta_example(), c("apat", "wk12", "ae_summary"))

}
