% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_run.R
\name{meta_run}
\alias{meta_run}
\title{Execute analysis based on the analysis plan.}
\usage{
meta_run(meta, i = NULL, ...)
}
\arguments{
\item{meta}{a \code{meta_adam} object.}

\item{i}{a vector of integer to indicate \code{i}th analysis in \code{meta$plan}.}

\item{...}{additional arguments transfer to \code{spec_call_program}.}
}
\description{
Execute analysis based on the analysis plan.
}
\examples{
meta <- meta_dummy()
ae_summary <- function(...) {
  paste("results of", deparse(match.call(), nlines = 1))
}
ae_specific <- function(...) {
  paste("results of", deparse(match.call(), nlines = 1))
}
meta_run(meta)
meta_run(meta, i = 2)
}
