% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metalite-package}
\alias{metalite}
\alias{metalite-package}
\title{metalite: ADaM Metadata Structure}
\description{
An R package to create metadata structure for ADaM data analysis and reporting in clinical trial development.
}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao1@merck.com}

Authors:
\itemize{
  \item Yilong Zhang
  \item Ruchitbhai Patel
  \item Madhusudhan Ginnaram
  \item Sarad Nepal
  \item Yujie Zhao
}

Other contributors:
\itemize{
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
