% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdata.R
\name{new_outdata}
\alias{new_outdata}
\title{Structure outdata class}
\usage{
new_outdata(x, env = globalenv())
}
\arguments{
\item{x}{A list, see Details.}

\item{env}{An environment.}
}
\description{
Structure outdata class
}
\details{
The list contains these elements:
\itemize{
\item \code{meta}
\item \code{population}
\item \code{observation}
\item \code{parameter}
\item \code{n}
\item \code{order}
\item \code{group}
\item \code{reference_group}
}
}
\examples{
meta <- meta_dummy()
x <- list(
  meta = meta_dummy(),
  population = "apat",
  observation = "wk12",
  parameter = "rel",
  n = data.frame(
    TRTA = c("Placebo", "Xanomeline Low Dose", "Xanomeline High Dose"),
    n = c(86, 84, 84)
  ),
  group = "TRTA",
  reference_group = 1,
  order = 1:3
)
metalite:::new_outdata(x)
}
