% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{omit_null}
\alias{omit_null}
\title{Omit NULL values in a list}
\usage{
omit_null(x)
}
\arguments{
\item{x}{a list.}
}
\description{
Omit NULL values in a list
}
\examples{
\dontrun{
  omit_null(list(a = 1, b = NULL))
}

}
