% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_adam.R
\name{print.meta_adam}
\alias{print.meta_adam}
\title{Print a meta data with its population, observation and analysis plans}
\usage{
\method{print}{meta_adam}(x, ...)
}
\arguments{
\item{x}{a object returned by \code{meta_adam}}

\item{...}{additional variables save to \code{print}}
}
\description{
Print a meta data with its population, observation and analysis plans
}
\examples{
meta_adam(observation = r2rtf::r2rtf_adae, population = r2rtf::r2rtf_adae) |> print()

}
