% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reset_label}
\alias{reset_label}
\title{Reset Dataset Label}
\usage{
reset_label(data, data_label)
}
\arguments{
\item{data}{a data frame}

\item{data_label}{a data frame with label}
}
\description{
Reset Dataset Label
}
