% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{rtf_mock_color}
\alias{rtf_mock_color}
\title{Enable color for mockup table}
\usage{
rtf_mock_color(text)
}
\arguments{
\item{text}{a string.}
}
\description{
The function will automatically replace text between \code{<} and \code{>} into blue.
}
\examples{
library(r2rtf)

head(iris) \%>\%
  r2rtf::rtf_title(rtf_mock_color("{week}")) \%>\%
  rtf_body() \%>\%
  rtf_encode() \%>\%
  write_rtf(file = paste0("~/ex-1.rtf"))
}
