% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_call_program}
\alias{spec_call_program}
\title{Specification for analysis call program}
\usage{
spec_call_program(meta, ...)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{...}{additional arguments used in all call programs.}
}
\description{
Specification for analysis call program
}
\examples{
library(r2rtf)
library(magrittr)
meta <- meta_dummy()

spec_call_program(meta)
spec_call_program(meta, data_source = "[Study CDISCpilot: adam-adsl; adae]")
}
