% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_filename}
\alias{spec_filename}
\title{Specification for analysis output filename}
\usage{
spec_filename(meta)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}
}
\description{
Specification for analysis output filename
}
\examples{
library(r2rtf)
meta <- meta_dummy()

spec_filename(meta)
}
