% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_title}
\alias{spec_title}
\title{Specification for analysis title}
\usage{
spec_title(meta)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}
}
\description{
Specification for analysis title
}
\examples{
library(r2rtf)
library(magrittr)
meta <- meta_dummy()
  
spec_title(meta)

}
