% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_adam_mapping}
\alias{update_adam_mapping}
\title{Update mapping rule in \code{adam_mapping}}
\usage{
update_adam_mapping(meta, name, ...)
}
\arguments{
\item{meta}{a \code{meta_adam} object.}

\item{name}{a vector of keywords}

\item{...}{additional variables to be added in the mapping rule among those keywords.}
}
\description{
Update mapping rule in \code{adam_mapping}
}
\examples{
meta <- meta_dummy()
meta <- update_adam_mapping(meta, names(meta$parameter), start_date = "ASTDT")
collect_adam_mapping(meta, "ser")
}
