% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_mapping.R
\name{validate_adam_mapping}
\alias{validate_adam_mapping}
\title{Check adam_mapping class}
\usage{
validate_adam_mapping(x)
}
\arguments{
\item{x}{an \code{adam_mapping} object}
}
\description{
Check adam_mapping class
}
\examples{
\dontrun{
validate_adam_mapping(new_adam_mapping(list(name = "apat")))
}

}
