% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdata.R
\name{validate_outdata}
\alias{validate_outdata}
\title{Validate outdata class}
\usage{
validate_outdata(x)
}
\arguments{
\item{x}{an \code{outdata} object}
}
\description{
Validate outdata class
}
\examples{
meta <- meta_dummy()
metalite:::validate_outdata(
  metalite:::outdata(
    meta = meta_dummy(),
    population = "apat",
    observation = "wk12",
    parameter = "rel",
    n = data.frame(
      TRTA = c("Placebo", "Xanomeline Low Dose", "Xanomeline High Dose"),
      n = c(86, 84, 84)
    ),
    group = "TRTA",
    reference_group = 1,
    order = 1:3
  )
)
}
