% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_adam_mapping}
\alias{collect_adam_mapping}
\title{Collect \code{adam_mapping} from \code{meta_adam} by \code{name}}
\usage{
collect_adam_mapping(meta, name)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{name}{A keyword value.}
}
\value{
An \code{adam_mapping} class object containing the definition of
the search variable in \code{name}.
}
\description{
Collect \code{adam_mapping} from \code{meta_adam} by \code{name}
}
\examples{
meta <- meta_example()
collect_adam_mapping(meta, "apat")
}
