% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_dataname}
\alias{collect_dataname}
\title{Collect specification for dataset name}
\usage{
collect_dataname(meta)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}
}
\description{
Collect specification for dataset name
}
\examples{
library(r2rtf)
library(magrittr)
meta <- meta_dummy()
  
collect_dataname(meta)

}
