% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_dataname}
\alias{collect_dataname}
\title{Collect specification for dataset name}
\usage{
collect_dataname(meta)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}
}
\value{
A vector of character strings containing the name of
the population/observation.
}
\description{
Collect specification for dataset name
}
\examples{
meta <- meta_example()
collect_dataname(meta)
}
