% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_observation_record}
\alias{collect_observation_record}
\title{Collect observation record from observation dataset}
\usage{
collect_observation_record(
  meta,
  population,
  observation,
  parameter,
  var = NULL
)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}

\item{var}{A character vector of additional variables to be displayed
in the output.}
}
\value{
A data frame of the observation dataset.
}
\description{
The key variables used in \code{id}, \code{group}, and \code{subset}
are displayed by default.
}
\examples{
meta <- meta_example()
collect_observation_record(meta, "apat", "wk12", "ser")
collect_observation_record(meta, "apat", "wk12", "ser", var = "AEDECOD")
}
