% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_population}
\alias{collect_population}
\title{Collect specification for population definition}
\usage{
collect_population(meta, population, observation, parameter)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}
}
\description{
Collect specification for population definition
}
\examples{
library(r2rtf)
library(magrittr)
meta <- meta_dummy()

collect_population(meta, "apat", "wk12", "ser")
}
