% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_population}
\alias{collect_population}
\title{Collect specification for population definition}
\usage{
collect_population(meta, population, observation = NULL, parameter = NULL)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}
}
\value{
A list covering the filter of population,
observation (if given) and parameter (if given).
}
\description{
Collect specification for population definition
}
\examples{
meta <- meta_example()
collect_population(meta, "apat")
collect_population(meta, "apat", "wk12")
collect_population(meta, "apat", "wk12", "ser")
}
