% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_population_id}
\alias{collect_population_id}
\title{Collect subject identifier information from population dataset}
\usage{
collect_population_id(meta, population)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}
}
\value{
A vector of patient ID within the population group.
}
\description{
Collect subject identifier information from population dataset
}
\examples{
meta <- meta_example()
head(collect_population_id(meta, "apat"))
}
