% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_population_index}
\alias{collect_population_index}
\title{Collect population record index from population dataset}
\usage{
collect_population_index(meta, population)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}
}
\value{
A vector of patient index within the population group.
}
\description{
Collect population record index from population dataset
}
\examples{
meta <- meta_example()
head(collect_population_index(meta, "apat"))
}
