% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_population_record}
\alias{collect_population_record}
\title{Collect population record from population dataset}
\usage{
collect_population_record(meta, population, var = NULL)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{var}{A character vector of additional variables to be displayed
in the output.}
}
\value{
A data frame containing the variables in the population dataset.
}
\description{
The key variables used in \code{id}, \code{group}, and \code{subset}
are displayed by default.
}
\examples{
meta <- meta_example()
head(collect_population_record(meta, "apat"))
head(collect_population_record(meta, "apat", var = "AGE"))
}
