% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect_title}
\alias{collect_title}
\title{Collect specification for title}
\usage{
collect_title(
  meta,
  population,
  observation,
  parameter,
  analysis,
  title_order = c("analysis", "observation", "population")
)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{population}{A character value of population term name.
The term name is used as key to link information.}

\item{observation}{A character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{A character value of parameter term name.
The term name is used as key to link information.}

\item{analysis}{A character value of analysis term name.
The term name is used as key to link information.}

\item{title_order}{A character vector to define the order of title
from each component.}
}
\value{
A vector of strings to compose the table captions.
}
\description{
Collect specification for title
}
\examples{
meta <- meta_example()
collect_title(meta, "apat", "wk12", "ser", "ae_summary")
collect_title(meta, "apat", "wk12", "ser", "ae_specific")
}
