% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{default_apply}
\alias{default_apply}
\title{Apply default values to ADaM mappings}
\usage{
default_apply(x)
}
\arguments{
\item{x}{An \code{adam_mapping} object.}
}
\value{
Similar to the input, but with the missing values
updated to the default values.
}
\description{
Apply default values to ADaM mappings
}
\examples{
default_apply(adam_mapping(name = "apat"))
}
