% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_analysis}
\alias{define_analysis}
\title{Define analysis function meta information for ADaM dataset}
\usage{
define_analysis(meta, name, ...)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{name}{a character value of term name. The term name is used as key to link information.}

\item{...}{additional variables.}
}
\description{
Define analysis function meta information for ADaM dataset
}
\examples{
meta_adam(
  population = r2rtf::r2rtf_adsl,
  observation = r2rtf::r2rtf_adae
) |>
  define_plan(plan = plan) |>
  define_analysis(
    name = "ae_summary",
    title = "Summary of Adverse Events"
  )
}
