% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_plan}
\alias{define_plan}
\title{Define analysis plan meta information for ADaM dataset}
\usage{
define_plan(meta, plan)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{plan}{A data frame for analysis plan.}
}
\value{
A metadata object with plans defined.
}
\description{
Define analysis plan meta information for ADaM dataset
}
\examples{
plan <- plan(
  analysis = "ae_summary", population = "apat",
  observation = c("wk12", "wk24"), parameter = "any;rel;ser"
)

meta_adam(
  population = r2rtf::r2rtf_adsl,
  observation = r2rtf::r2rtf_adae
) |>
  define_plan(plan)
}
