% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.R
\name{define_population}
\alias{define_population}
\title{Define analysis population meta information for ADaM dataset}
\usage{
define_population(
  meta,
  name,
  id = "USUBJID",
  group = NULL,
  var = NULL,
  subset = NULL,
  label = NULL,
  ...
)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}

\item{name}{a character value of term name. The term name is used as key to link information.}

\item{id}{a character value of subject id variable name in an ADaM dataset.}

\item{group}{a character vector of group variable names in an ADaM dataset.}

\item{var}{a character vector of useful variable names in an ADaM dataset.}

\item{subset}{an expression to identify analysis records. Refer \code{subset()} function.}

\item{label}{a character value of analysis label.}

\item{...}{additional variables.}
}
\description{
Define analysis population meta information for ADaM dataset
}
\examples{
library(r2rtf)
plan <- plan(analysis = "ae_summary", population = "apat", 
observation = c("wk12", "wk24"), parameter = "any;rel;ser")

meta_adam(
  population = r2rtf::r2rtf_adsl,
  observation = r2rtf::r2rtf_adae
) |>
  define_plan(plan) |>
  define_population(name = "apat")
}
