% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_adam.R
\name{meta_adam}
\alias{meta_adam}
\title{Create a metadata representation for ADaM data analysis}
\usage{
meta_adam(observation, population = observation)
}
\arguments{
\item{observation}{a data frame for observation level data}

\item{population}{a data frame for population level data. Default is the same as \code{observation}}
}
\value{
a metadata initialization with observation and population defined.
}
\description{
Create a metadata representation for ADaM data analysis
}
\examples{
meta_adam(observation = r2rtf::r2rtf_adae, population = r2rtf::r2rtf_adae)
}
