% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_add_total.R
\name{meta_add_total}
\alias{meta_add_total}
\title{Add duplicate data to enable a total group.}
\usage{
meta_add_total(meta, total = "Total")
}
\arguments{
\item{meta}{a metalite object}

\item{total}{a character value of total group name.}
}
\description{
Add duplicate data to enable a total group.
}
