% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_dummy_exploration.R
\name{meta_dummy_exploration}
\alias{meta_dummy_exploration}
\title{Create a data exploration \code{meta_adam} object}
\usage{
meta_dummy_exploration(
  data,
  group,
  name = "ase",
  subset = NULL,
  label = "All Subjects Enrolled"
)
}
\arguments{
\item{data}{a data frame}

\item{group}{a character vector of group variable names in an ADaM dataset.}

\item{name}{a character value of term name. The term name is used as key to link information.}

\item{subset}{an expression to identify analysis records. Refer \code{subset()} function.}

\item{label}{a character value of analysis label.}
}
\description{
Create a data exploration \code{meta_adam} object
}
\examples{
meta <- meta_dummy_exploration(r2rtf::r2rtf_adsl, group = "TRT01A")
collect_n(meta, "ase", "AGE")
collect_n(meta, "ase", "SEX")

}
