% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_run.R
\name{meta_run}
\alias{meta_run}
\title{Execute analysis based on the analysis plan}
\usage{
meta_run(meta, i = NULL, ...)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{i}{A vector of integers to indicate \code{i}-th analysis in \code{meta$plan}.}

\item{...}{Additional arguments passed to \verb{[spec_call_program()]}.}
}
\value{
Executed analysis based on the analysis plan.
}
\description{
Execute analysis based on the analysis plan
}
\examples{
meta <- meta_example()
ae_summary <- function(...) {
  paste("results of", deparse(match.call(), nlines = 1))
}
ae_specific <- function(...) {
  paste("results of", deparse(match.call(), nlines = 1))
}
meta_run(meta)
meta_run(meta, i = 2)
}
