% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_split.R
\name{meta_split}
\alias{meta_split}
\title{Devide Meta into Groups}
\usage{
meta_split(meta, by)
}
\arguments{
\item{meta}{a \code{meta_adam} object.}

\item{by}{a character variable name both in
population level and observation level data of a meta object.}
}
\description{
Devide Meta into Groups
}
\examples{
meta_dummy() |>
  meta_split("RACE")
}
