% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metalite-package}
\alias{metalite}
\alias{metalite-package}
\title{metalite: ADaM Metadata Structure}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create metadata structure for ADaM data analysis and reporting in clinical trial development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/metalite/}
  \item \url{https://github.com/Merck/metalite}
  \item Report bugs at \url{https://github.com/Merck/metalite/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao1@merck.com}

Authors:
\itemize{
  \item Yilong Zhang
  \item Ruchitbhai Patel
  \item Madhusudhan Ginnaram
  \item Sarad Nepal
  \item Yujie Zhao
}

Other contributors:
\itemize{
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
