% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metalite-package}
\alias{metalite}
\alias{metalite-package}
\title{metalite: ADaM metadata structure}
\description{
An R package to create metadata structure for ADaM data analysis and reporting in clinical trial development.
}
\author{
\strong{Maintainer}: Nan Xiao \email{nan.xiao1@merck.com}

Other contributors:
\itemize{
  \item BARDS, Merck & Co., Inc. [copyright holder]
}

}
\keyword{internal}
