% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metalite-package}
\alias{metalite}
\alias{metalite-package}
\title{metalite: ADaM Metadata Structure}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A metadata structure for clinical data analysis and reporting based on Analysis Data Model (ADaM) datasets. The package simplifies clinical analysis and reporting tool development by defining standardized inputs, outputs, and workflow. The package can be used to create analysis and reporting planning grid, mock table, and validated analysis and reporting results based on consistent inputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://merck.github.io/metalite/}
  \item \url{https://github.com/Merck/metalite}
  \item Report bugs at \url{https://github.com/Merck/metalite/issues}
}

}
\author{
\strong{Maintainer}: Yujie Zhao \email{yujie.zhao@merck.com}

Authors:
\itemize{
  \item Yilong Zhang
  \item Nan Xiao
}

Other contributors:
\itemize{
  \item Benjamin Wang [contributor]
  \item Brian Lang [contributor]
  \item Howard Baek [contributor]
  \item Ruchitbhai Patel [contributor]
  \item Madhusudhan Ginnaram [contributor]
  \item Sarad Nepal [contributor]
  \item Venkatesh Burla [contributor]
  \item Merck Sharp & Dohme Corp [copyright holder]
}

}
\keyword{internal}
