% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdata.R
\name{new_outdata}
\alias{new_outdata}
\title{Validate outdata class}
\usage{
new_outdata(x, env = globalenv())
}
\arguments{
\item{x}{an \code{outdata} object}

\item{env}{an environment}
}
\description{
Validate outdata class
}
