% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdata.R
\name{new_outdata}
\alias{new_outdata}
\title{Structure outdata class}
\usage{
new_outdata(x, env = globalenv())
}
\arguments{
\item{x}{an \code{outdata} object}

\item{env}{an environment}
}
\description{
Structure outdata class
}
\examples{
library(dplyr)
meta <- meta_dummy()
x <- list(meta = meta_dummy(), 
          population = "apat", 
          observation = "wk12", 
          parameter = "rel", 
          n = meta$data_population \%>\% group_by(TRTA) \%>\% summarize(n = n()), 
          group = "TRTA", 
          reference_group = 1, 
          order = 1:3)
metalite:::new_outdata(x)
}
