% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdata.R
\name{new_outdata}
\alias{new_outdata}
\title{Structure outdata class}
\usage{
new_outdata(x, env = globalenv())
}
\arguments{
\item{x}{an list object containing elements: meta, population, observation, parameter, n, order, group, and reference_group}

\item{env}{an environment}
}
\description{
Structure outdata class
}
\examples{
meta <- meta_dummy()
x <- list(
  meta = meta_dummy(),
  population = "apat",
  observation = "wk12",
  parameter = "rel",
  n = data.frame(
    TRTA = c("Placebo", "Xanomeline Low Dose", "Xanomeline High Dose"),
    n = c(86, 84, 84)
  ),
  group = "TRTA",
  reference_group = 1,
  order = 1:3
)
metalite:::new_outdata(x)
}
