% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{new_plan}
\alias{new_plan}
\title{Create a analysis plan from all combination of variables}
\usage{
new_plan(analysis, population, observation, parameter, mock = 1, ...)
}
\arguments{
\item{analysis}{a character value of analysis term name.
The term name is used as key to link information.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}

\item{mock}{a numeric value of mock table number.}

\item{...}{additional arguments}
}
\description{
This function is a wrapper of \code{expand.grid}
}
\examples{
metalite:::new_plan(analysis = "ae_specific", 
                    population = "apat", 
                    observation = c("wk12", "wk24"), 
                    parameter = c("any", "rel", "ser"))
}
