% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{plan}
\alias{plan}
\title{Create a analysis plan from all combination of variables}
\usage{
plan(analysis, population, observation, parameter, mock = 1, ...)
}
\arguments{
\item{analysis}{a character value of analysis term name.
The term name is used as key to link information.}

\item{population}{a character value of population term name.
The term name is used as key to link information.}

\item{observation}{a character value of observation term name.
The term name is used as key to link information.}

\item{parameter}{a character value of parameter term name.
The term name is used as key to link information.}

\item{mock}{a numeric value of mock table number.}

\item{...}{additional arguments}
}
\value{
a data.frame showing the analysis plans
}
\description{
This function is a wrapper of \code{expand.grid}
}
\examples{

# example 1
# create an analysis plan of AE summary
# with any AE, drug-related AE and serious AE
plan(
  analysis = "ae_summary",
  population = "apat",
  observation = c("wk12", "wk24"),
  parameter = "any;rel;ser"
)

# example 2
# create an analysis plan of AE specific
# with any AE, drug-related AE and serious AE
plan(
  analysis = "ae_specific",
  population = "apat",
  observation = c("wk12", "wk24"),
  parameter = c("any", "rel", "ser")
)
}
