% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_adam.R
\name{print.meta_adam}
\alias{print.meta_adam}
\title{Print a metadata object with its population, observation, and analysis plans}
\usage{
\method{print}{meta_adam}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link[=meta_adam]{meta_adam()}}.}

\item{...}{Additional parameters for \code{\link[=print]{print()}} (not used).}
}
\value{
A printed summary of the metadata.
}
\description{
Print a metadata object with its population, observation, and analysis plans
}
\examples{
meta_adam(observation = r2rtf::r2rtf_adae, population = r2rtf::r2rtf_adae) |> print()
}
