% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{reset_label}
\alias{reset_label}
\title{Reset dataset label}
\usage{
reset_label(data, data_label)
}
\arguments{
\item{data}{A data frame.}

\item{data_label}{A data frame with label.}
}
\description{
Reset dataset label
}
