% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtf.R
\name{rtf_assemble}
\alias{rtf_assemble}
\title{Assemble RTF files into one single RTF file}
\usage{
rtf_assemble(path)
}
\arguments{
\item{path}{a vector of RTF file path}
}
\description{
Assemble RTF files into one single RTF file
}
\examples{
library(r2rtf)

head(iris) \%>\%
  rtf_body() \%>\%
  rtf_encode() \%>\%
  write_rtf(file = paste0(tempdir(), "/ex-1.rtf"))

tail(iris) \%>\%
  rtf_body() \%>\%
  rtf_encode() \%>\%
  write_rtf(file = paste0(tempdir(), "/ex-2.rtf"))

rtf_assemble(c(paste0(tempdir(), "/ex-1.rtf"), paste0(tempdir(), "/ex-2.rtf")))
}
