% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_analysis_population}
\alias{spec_analysis_population}
\title{Specification for population definition}
\usage{
spec_analysis_population(meta)
}
\arguments{
\item{meta}{an \code{meta_adam} object.}
}
\value{
a vector of character showing the populations used
in the order of the analysis plans
}
\description{
Specification for population definition
}
\examples{
library(r2rtf)
meta <- meta_dummy()

spec_analysis_population(meta)
}
