% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_call_program}
\alias{spec_call_program}
\title{Specification for analysis call program}
\usage{
spec_call_program(meta, ...)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{...}{Additional arguments used in all call programs.}
}
\value{
A vector of character strings containing the call program
in the order of the analysis plans.
}
\description{
Specification for analysis call program
}
\examples{
meta <- meta_example()
spec_call_program(meta)
spec_call_program(meta, data_source = "[Study CDISCpilot: adam-adsl; adae]")
}
