% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_filename}
\alias{spec_filename}
\title{Specification for analysis output filename}
\usage{
spec_filename(meta)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}
}
\value{
A vector of character strings containing the RTF file names.
}
\description{
Specification for analysis output filename
}
\examples{
meta <- meta_example()
spec_filename(meta)
}
