% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_title}
\alias{spec_title}
\title{Specification for analysis title}
\usage{
spec_title(meta)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}
}
\value{
A vector of character strings containing the table captions
in the order of the analysis plans.
}
\description{
Specification for analysis title
}
\examples{
meta <- meta_example()
spec_title(meta)
}
