% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_adam_mapping}
\alias{update_adam_mapping}
\title{Update mapping rule in \code{adam_mapping}}
\usage{
update_adam_mapping(meta, name, ...)
}
\arguments{
\item{meta}{A \code{meta_adam} object.}

\item{name}{A vector of keywords.}

\item{...}{Additional variables to be added in the mapping rule
among those keywords.}
}
\value{
A metadata object with the input updated.
}
\description{
Update mapping rule in \code{adam_mapping}
}
\examples{
meta <- meta_example()
meta <- update_adam_mapping(meta, names(meta$parameter), start_date = "ASTDT")
collect_adam_mapping(meta, "ser")
}
